Improve the dataset builder
The builders for the datasets are now definex in interfaces. The Dataset class
creates anonymous instances. The builders are divided in three types, one for
the information that are not impacting the order of the values, one for the
dimensions and one for the values.

The Dataset is now made abstract to allow custom implementations of the
values collection. This allowed to use a cartesian list from Guave to compute
the posisions in the values collection.